fix: build/type issue with DataGenerator
- Move around types to exclude mocks directory
- Remove import of seedrandom from dist files
- Add default rng function for DataGenerator
