+++ tsconfig.lib.json
@@ -5,5 +5,5 @@
   },
   "extends": "./tsconfig",
   "include": ["src/**/*"],
-  "exclude": ["**/*.test.*", "**/__mocks__"]
+  "exclude": ["**/*.test.*", "**/__mocks__", "src/mocks/**/*"]
 }
